requirements.txt
Django==5.0.1
django-allauth==0.60.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-environ==0.11.2
django-filter==23.5
django-imagekit==5.0.0
django-redis==5.4.0
djangorestframework==3.14.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-magic==0.4.27
redis==5.0.1
requests==2.31.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
celery==5.3.6
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
bcrypt==4.1.2
django-storages==1.14.2
boto3==1.34.14
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-ckeditor==6.7.0
django-compressor==4.4
django-phonenumber-field[phonenumbers]==7.3.0
django-otp==1.3.0
qrcode==7.4.2
django-countries==7.5.1
django-timezone-field==6.1.0
django-modeltranslation==0.18.11
django-crontab==0.7.1
django-maintenance-mode==0.21.0
django-admin-interface==0.28.3
django-import-export==3.3.6
django-simple-history==3.4.0
django-activity-stream==2.0.0
django-notifications-hq==1.8.3
django-push-notifications==3.0.0
django-ratelimit==4.1.0
django-recaptcha==3.0.0
django-robots==6.1
django-user-agents==0.4.0
django-widget-tweaks==1.5.0
django-tinymce==3.7.1
django-summernote==0.8.20.0
django-taggit==5.0.1